Update template layout elements, inject custom styles
Wrap the price and last updated elements in a <div> to ensure that
they clear floats or don't together.

Injecting styles using an inline stylesheet in this way is something
of a workaround to deal with the fact that most Onebox styles are
actually implemented within the main Discourse project, not within
the Onebox extension. While inefficient in terms of data, it's a very
reasonable compromise when faced with the possibility of needing to
maintain a fork of the main Discourse project.

Also, since it wasn't possible to use data URLs as background images
in the injected styles, the header icon is hotlinked from the following
location: http://www.thewinesociety.com/images/fb/tws-logo.png.
While this represents some risk if the image should be moved or removed
in the future, it's more efficient in terms of data transfer. Even so,
this should be considered a temporary workaround until a better solution
is found.